backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
SQLAlchemy==2.0.40
pydantic-settings==2.8.1
email-validator==2.2.0
python-dotenv==1.1.0
psycopg[binary]==3.2.6
httpx==0.28.1
